feat: made a bot feat: added delay to bot's shots fix: fixed bugs

+++ src/game_server/controller/controller.ts
@@ -1,4 +1,5 @@
 import { DB } from "../db";
+import { Bot } from "./bot";
 import { Game } from "./game";
 
 class Session {
@@ -12,6 +13,7 @@ class Sessions {
         const lastId = this.sessions[this.sessions.length - 1]?.id;
         const id = lastId !== undefined ? lastId + 1 : 0;
         this.sessions.push(new Session(id, userIndex));
+
         return id;
     }
 
@@ -26,6 +28,7 @@ class Sessions {
 
 export class Controller {
     sessions = new Sessions();
+    bots = new Map<number, Bot>();
 
     constructor(private db: DB) { }
 
@@ -49,12 +52,12 @@ export class Controller {
                         error: true,
                         errorText: 'Wrong name and password for existing user',
                     };
-                    responses.push(this.makeResponse('reg', data));
+                    responses.push(this.makeResponse('reg', data, []));
                     break;
                 }
                 const sessionId = this.sessions.addSession(user.index);
                 user.index = sessionId;
-                responses.push(this.makeResponse('reg', user));
+                responses.push(this.makeResponse('reg', user, []));
                 responses.push(this.makeResponse('update_room', this.getRooms(), 'broadcast'));
                 responses.push(this.makeResponse('update_winners', this.db.getWinners(), 'broadcast'));
                 break;
@@ -82,6 +85,9 @@ export class Controller {
 
             case 'add_ships':
                 game = this.db.games[dataObj.data.gameId];
+                if (!game) {
+                    break;
+                }
                 result = game.addShips(dataObj.data.indexPlayer, dataObj.data.ships);
                 if (!result) {
                     break;
@@ -98,7 +104,11 @@ export class Controller {
             case 'randomAttack':
                 const { gameId, x, y, indexPlayer } = dataObj.data;
                 game = this.db.games[gameId];
-                result = dataObj.type === 'attack' ?  game.attack(indexPlayer, x, y) : game.randomAttack(indexPlayer);
+                if (!game) {
+                    break;
+                }
+
+                result = dataObj.type === 'attack' ? game.attack(indexPlayer, x, y) : game.randomAttack(indexPlayer);
                 if (!result) {
                     break;
                 }
@@ -106,16 +116,53 @@ export class Controller {
                 players = game.getPlayers();
                 for (const cell of result.cells) {
                     responses.push(this.makeResponse('attack', cell, players));
-                }
-                
-                if (result.win) {
-                    this.db.addWinner(this.sessions.getUserIndex(userIndex));
-                    responses.push(this.makeResponse('update_winners', this.db.getWinners(), 'broadcast'));
-                    responses.push(this.makeResponse('finish', result.win, players));
-                    this.db.deleteGame(game.id);
-                } else {
                     responses.push(this.makeResponse('turn', result.turn, players));
                 }
+
+                if (result.win) {
+                    const winPlayer = result.win.winPlayer;
+                    // if (winPlayer !== Bot.id) {
+                        this.db.addWinner(this.sessions.getUserIndex(winPlayer));
+                        responses.push(this.makeResponse('update_winners', this.db.getWinners(), 'broadcast'));
+                    // }
+                    responses.push(this.makeResponse('finish', result.win, players));
+                    if (game.gameWithBot) {
+                        this.bots.delete(game.id);
+                    }
+                    this.db.deleteGame(game.id);
+                } else {
+                    if (result.turn.currentPlayer === Bot.id) {
+                        const bot = this.bots.get(game.id);
+                        result = bot.attack();
+                        for (let i = 0; i < result.cells.length; i += 1) {
+                            const cell = result.cells[i];
+                            const turn = i === result.cells.length - 1 ? userIndex : Bot.id;
+
+                            const status = cell.status;
+                            let delay = false;
+                            if (status === 'shot' || status === 'killed' || (status ==='miss' && i === result.cells.length - 1)) {
+                                delay = true;
+                            }
+                    
+                            responses.push(this.makeResponse('attack', cell, players, delay));
+                            responses.push(this.makeResponse('turn', { currentPlayer: turn }, players));
+                        }
+                        if (result.win) {
+                            responses.push(this.makeResponse('finish', result.win, players));
+                            this.bots.delete(game.id);
+                            this.db.deleteGame(game.id);
+                        }
+                    }
+                }
+                break;
+
+            case 'single_play':
+                const gameData = this.db.createGame(userIndex);
+                game = this.db.getGame(gameData[0].idGame);
+                const bot = new Bot(game);
+                this.bots.set(game.id, bot);
+                responses.push(this.makeResponse('create_game', gameData[0], [userIndex]));
+                responses.push(this.makeResponse('update_room', this.getRooms(), 'broadcast'));
                 break;
         }
 
@@ -127,11 +174,17 @@ export class Controller {
         for (const room of rooms) {
             (room as any).roomUsers[0].name = this.db.getUser(this.sessions.getUserIndex(room.roomUsers[0].index)).name;
         }
+
         return rooms;
     }
 
-    private makeResponse(type: string, data: any, receivers?: number[] | 'broadcast') {
+    private makeResponse(type: string, data: any, receivers: number[] | 'broadcast', delay?: boolean) {
+        if (Array.isArray(receivers) && receivers.includes(Bot.id)) {
+            receivers.splice(receivers.indexOf(Bot.id), 1);
+        }
+
         return {
+            delay: !!delay,
             receivers,
             payload: {
                 type,
@@ -148,8 +201,14 @@ export class Controller {
             return [this.makeResponse('update_room', this.getRooms(), 'broadcast')];
         }
 
-        const game = this.db.games.find((game) => game.getPlayers().find((player) => player === userIndex));
+        const game = this.db.games.find((game) => game.getPlayers().includes(userIndex));
         if (game !== undefined) {
+            if (game.gameWithBot) {
+                this.db.deleteGame(game.id);
+                this.bots.delete(game.id);
+                return [];
+            }
+
             const responses: any[] = [];
             const opponent = game.getOpponent(userIndex);
             this.db.addWinner(this.sessions.getUserIndex(opponent));
